games/tic-tac-toe/gateway-service/requirements.txt
websockets==12.0
kafka-python==2.0.2


